+++ packages/nhsuk-frontend/src/nhsuk/components/tag/macro-options.mjs
@@ -8,6 +8,11 @@ export const name = 'Tag'
  * @satisfies {{ [param: string]: MacroParam }}
  */
 export const params = {
+  id: {
+    type: 'string',
+    required: false,
+    description: 'The ID of the tag.'
+  },
   text: {
     type: 'string',
     required: true,