+++ CHANGELOG.md
@@ -31,6 +31,10 @@ If you are not using Nunjucks macros, update your HTML markup to:
 
 You can now use the action link component on dark backgrounds by using the `.nhsuk-action-link--reverse` class. Added in [pull request #1542: Add reverse action link modifiers and styles](https://github.com/nhsuk/nhsuk-frontend/pull/1542).
 
+:wrench: **Fixes**
+
+- [#1536: Add support for id param to all components](https://github.com/nhsuk/nhsuk-frontend/pull/1536)
+
 ## 10.0.0-internal.3 - 15 August 2025
 
 :new: **New features**